Base class for drifter Ensemble
